third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/116-expected.txt

This is a testharness.js-based test.
FAIL scheduler: adding script to head of frameset document assert_array_equals: expected property 0 to be "document.body: <FRAMESET>" but got "document.body: <BODY>" (expected array ["document.body: <FRAMESET>"] got ["document.body: <BODY>"])
Harness: the test ran to completion.
